move sidebar components to individual files
