+++ components/Navigation/Sidebar/Sidebar.js
@@ -1,11 +1,11 @@
 import Link from "next/link"
-import { useRouter } from "next/router"
 import { Tooltip } from "react-tooltip"
 
 import Toggle from "@components/Toggle"
 import { Text } from "@radix-ui/themes"
-import { usePathname } from "next/navigation"
-import { BoltIcon, HomeIcon, SquaresPlusIcon, TableCellsIcon } from "@heroicons/react/20/solid"
+import { HomeIcon, SquaresPlusIcon, TableCellsIcon } from "@heroicons/react/20/solid"
+import { Item } from "./SidebarComponents/SidebarNavItem"
+import { SidebarFooter } from "./SidebarComponents/SidebarFooter"
 
 const navItems = [
   {
@@ -97,67 +97,3 @@ export default function Sidebar({ onLogOut, embedMode, setEmbedMode, devMode, se
     </div>
   )
 }
-
-function Item({ name, nested, href, icon: Icon }) {
-  const pathname = usePathname()
-  const isActive = pathname === href
-
-  return (
-    <Link
-      href={href}
-      className={`flex cursor-pointer flex-row items-center gap-2 rounded px-2 py-2 text-sm font-medium leading-none   ${
-        isActive
-          ? "bg-slate-200 text-slate-900 "
-          : "text-slate-600 hover:bg-slate-200/50 hover:text-slate-700"
-      }`}
-    >
-      {Icon && <Icon className={`h-4 ${isActive ? "text-slate-900" : ""}`} />}
-      <Text>{name}</Text>
-    </Link>
-  )
-}
-
-const SidebarFooter = ({ onLogOut, embedMode, setEmbedMode, devMode, setDevMode }) => (
-  <div className="flex flex-col items-center justify-between md:mt-auto md:w-full">
-    <div className="mb-3 hidden md:flex md:w-full md:flex-col md:gap-3">
-      <div className="flex items-center">
-        <span className="px-2">Embed</span>
-        <a id="embed-mode">
-          <Toggle checked={embedMode} onChange={() => setEmbedMode((prevCheck) => !prevCheck)} />
-        </a>
-        <Tooltip anchorSelect="#embed-mode">
-          Show Census UI flows embedded directly in the application, or redirect to the flows.
-        </Tooltip>
-      </div>
-      <div className="flex items-center">
-        <span className="px-2">Dev Mode</span>
-        <a id="dev-mode">
-          <Toggle checked={devMode} onChange={() => setDevMode((prevCheck) => !prevCheck)} />
-        </a>
-        <Tooltip anchorSelect="#dev-mode">
-          Show tooltips of the Census API requests being made when interacting with the UX.
-        </Tooltip>
-      </div>
-    </div>
-    <div className="flex items-center justify-between md:mt-auto md:w-full">
-      <button onClick={onLogOut} className="flex items-center gap-2 text-sm md:gap-3 md:text-2xl">
-        <i className="fa-solid fa-right-from-bracket mb-2 ml-2 md:mb-1" />
-        <p className="hidden md:block">Log out</p>
-      </button>
-      <div className="hidden items-center gap-4 text-2xl md:flex">
-        <Link
-          className="flex items-center gap-2 text-sm md:gap-3 md:text-2xl"
-          href="https://developers.getcensus.com/getting-started/introduction"
-        >
-          <i className="fa-solid fa-book" />
-        </Link>
-        <Link
-          className="flex items-center gap-2 text-sm md:gap-3 md:text-2xl"
-          href="https://github.com/sutrolabs/census-embedded-demo"
-        >
-          <i className="fa-brands fa-github" />
-        </Link>
-      </div>
-    </div>
-  </div>
-)
